+++ Detection/src/main/java/SignalProcessing/Effects/IIR_Filter.java
@@ -15,8 +15,8 @@ public class IIR_Filter implements IEffect
 {
     private IIR filter = null;
     private int max_chunk_size = 1024;
-    private final static double[] identity_FIR_coeffs = { 1 };
-    private double progress = 0;
+    private final static float[] identity_FIR_coeffs = { 1 };
+    private float progress = 0;
 
     @Override
     public void apply( IAudioDataSource dataSource, IAudioDataSource dataDest, Interval interval ) throws DataSourceException
@@ -36,11 +36,11 @@ public class IIR_Filter implements IEffect
         final boolean isInPlace = ( dataDest == dataSource );
         int buf1_size = filter.getFf_coeff_nr();
         int buf2_size = filter.getFb_coeff_nr() - 1;
-        double[][] prev_left_samples = null;
-        double[][] regression_buffer = new double[ dataSource.get_channel_number() ][ buf2_size ];
+        float[][] prev_left_samples = null;
+        float[][] regression_buffer = new float[ dataSource.get_channel_number() ][ buf2_size ];
         if( isInPlace )
         {
-            prev_left_samples = new double[ dataSource.get_channel_number() ][ buf1_size ];
+            prev_left_samples = new float[ dataSource.get_channel_number() ][ buf1_size ];
         }
 
         int temp_len;
@@ -89,7 +89,7 @@ public class IIR_Filter implements IEffect
         }
         dataDest.put_samples( win );
         i += applying_range.get_length();
-        progress = 1.0 * ( i - interval.l ) / interval.get_length();
+        progress = 1.0f * ( i - interval.l ) / interval.get_length();
 
         for( ; i < interval.r; )
         {
@@ -134,12 +134,12 @@ public class IIR_Filter implements IEffect
             }
             dataDest.put_samples( win );
             i += applying_range.get_length();
-            progress = 1.0 * ( i - interval.l ) / interval.get_length();
+            progress = 1.0f * ( i - interval.l ) / interval.get_length();
         }
     }
 
     @Override
-    public double getProgress()
+    public float getProgress()
     {
         return progress;
     }
